requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
streamlit==1.26.0
scikit-learn==1.3.2
joblib==1.3.2
requests==2.31.0
pydantic==1.10.13
python-multipart==0.0.6
sentence-transformers==2.2.2
qdrant-client==1.7.0
easyocr==1.6.2
Pillow==9.5.0
